refactor: make legilimens as a library
